gf : correction of matrix_stack with adding
- For +=,-= when it is a scalar, the type of the
"identity" array may be wrong e.g.
in demo a double. this would lead to a array of double,
which matrix_stack cython does not know how to
add to an array of complex.

+++ pytriqs/gf/local/gf_generic.py
@@ -161,7 +161,7 @@ class GfGeneric:
             MatrixStack(self.data).add(arg)
             t[0][:,:] += arg
         elif descriptors.is_scalar(arg): # just a scalar
-            arg = arg*numpy.identity(self.N1)
+            arg = arg*numpy.identity(self.N1,dtype = self.data.dtype )
             MatrixStack(self.data).add(arg)
             t[0][:,:] += arg
         else:
@@ -185,7 +185,7 @@ class GfGeneric:
             MatrixStack(self.data).sub(arg)
             t[0][:,:] -= arg
         elif descriptors.is_scalar(arg): # just a scalar
-            arg = arg*numpy.identity(self.N1)
+            arg = arg*numpy.identity(self.N1, dtype = self.data.dtype )
             MatrixStack(self.data).sub(arg)
             t[0][:,:] -= arg
         else:
